requirements.txt
pyspark==3.5.2
psycopg2-binary==2.9.10
pandas==2.2.3
pyarrow==18.1.0